refactor classes into packages
