+++ src/main/java/net/coderbot/iris/pipeline/transform/parameter/CompositeParameters.java
@@ -1,13 +1,13 @@
-package net.coderbot.iris.pipeline.transform;
+package net.coderbot.iris.pipeline.transform.parameter;
 
 import it.unimi.dsi.fastutil.objects.Object2ObjectMap;
-import net.coderbot.iris.gbuffer_overrides.matching.InputAvailability;
 import net.coderbot.iris.gl.blending.AlphaTest;
 import net.coderbot.iris.gl.texture.TextureType;
 import net.coderbot.iris.helpers.Tri;
+import net.coderbot.iris.pipeline.transform.Patch;
 import net.coderbot.iris.shaderpack.texture.TextureStage;
 
-class CompositeParameters extends Parameters {
+public class CompositeParameters extends Parameters {
 	public final TextureStage stage;
 	private final Object2ObjectMap<Tri<String, TextureType, TextureStage>, String> textureMap;
