+++ src/main/java/net/coderbot/iris/pipeline/transform/transformer/CommonTransformer.java
@@ -1,4 +1,4 @@
-package net.coderbot.iris.pipeline.transform;
+package net.coderbot.iris.pipeline.transform.transformer;
 
 import java.util.ArrayList;
 import java.util.HashSet;
@@ -32,8 +32,8 @@ import io.github.douira.glsl_transformer.util.Type;
 import net.coderbot.iris.Iris;
 import net.coderbot.iris.gl.blending.AlphaTest;
 import net.coderbot.iris.gl.shader.ShaderType;
+import net.coderbot.iris.pipeline.transform.parameter.Parameters;
 
-// Order fixed
 public class CommonTransformer {
 	public static final AutoHintedMatcher<Expression> glTextureMatrix0 = new AutoHintedMatcher<>(
 			"gl_TextureMatrix[0]", Matcher.expressionPattern);
