+++ src/main/java/net/coderbot/iris/pipeline/transform/transformer/TextureTransformer.java
@@ -1,4 +1,4 @@
-package net.coderbot.iris.pipeline.transform;
+package net.coderbot.iris.pipeline.transform.transformer;
 
 import io.github.douira.glsl_transformer.ast.node.Identifier;
 import io.github.douira.glsl_transformer.ast.node.TranslationUnit;
@@ -12,7 +12,7 @@ import net.coderbot.iris.gl.texture.TextureType;
 import net.coderbot.iris.helpers.Tri;
 import net.coderbot.iris.shaderpack.texture.TextureStage;
 
-class TextureTransformer {
+public class TextureTransformer {
 	public static void transform(
 			ASTParser t,
 			TranslationUnit tree,
